+++ MMO RPG/Game.h
@@ -18,7 +18,7 @@ public:
 	void Main();
 
 	void KeyPressed(const sf::Keyboard::Key key);
-	void MouseClicked(const sf::Vector2i position);
+	void MouseClicked(const sf::Vector2f position);
 
 private:
 	void UpdateModel();
